Added logic for rook moves, may need to be optmized

+++ ChessDotNet/BoardFactory.cs
@@ -87,7 +87,7 @@ namespace ChessDotNet
             return board;
         }
 
-        public ulong PiecesToBitBoard(IEnumerable<byte> pieces)
+        public ulong PiecesToBitBoard(IEnumerable<int> pieces)
         {
             var board = 0UL;
             foreach (var piece in pieces)
